Add advanced branch name mapping features
