+++ jira_branches/jira/ticket/ticket_info.py
@@ -1,0 +1,5 @@
+class TicketInfo:
+    def __init__(self, ticket_key, ticket_summary, ticket_type):
+        self.key = ticket_key
+        self.summary = ticket_summary
+        self.type = ticket_type
